# WIP Remove a chunk buffer in parseChunk