@wangkuiyi 2019-07-22
Use https://github.com/wangkuiyi/recordio/pull/57 instead.